refactor: MVC패턴에 맞춰 각 패키지로 이동
