+++ src/main/java/racingcar/Application.java
@@ -1,12 +1,10 @@
 package racingcar;
 
-import camp.nextstep.edu.missionutils.Console;
+import racingcar.controller.GameContoller;
 
 public class Application {
     public static void main(String[] args) {
-        // TODO: 프로그램 구현
-        //
-        Game game = new Game();
-        game.run();
+        GameContoller gameContoller = new GameContoller();
+        gameContoller.run();
     }
 }
