+++ src/main/java/racingcar/view/InputView.java
@@ -1,14 +1,14 @@
-package racingcar.io;
+package racingcar.view;
+
 
 import camp.nextstep.edu.missionutils.Console;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 
 public class InputView {
 
-    private static final int MAX_PLAYER_NAME_LENGTH=5;
+    private static final int MAX_PLAYER_NAME_LENGTH = 5;
 
     public List<String> readPlayerNameList() {
         System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
@@ -27,16 +27,25 @@ public class InputView {
         System.out.println("시도할 횟수는 몇 회인가요?");
         final String userInput=Console.readLine();
 
+        turn = parseInt(userInput);
+        checkPositive(turn);
+        return turn;
+    }
+
+    private void checkPositive(int turn) {
+        if(turn <=0){
+            throw new IllegalArgumentException("0이하의 값은 허용되지 않습니다.");
+        }
+    }
+
+    private int parseInt(String userInput) {
         try {
-            turn=Integer.parseInt(userInput);
+            return Integer.parseInt(userInput);
         }catch (NumberFormatException e){
             throw new IllegalArgumentException("숫자이외의 값은 허용되지 않습니다.");
         }
-        if(turn<=0){
-            throw new IllegalArgumentException("0이하의 값은 허용되지 않습니다.");
-        }
-        return turn;
     }
+
 
     private void checkDuplicate(List<String> playerNames) {
         Set<String> uniquePlayerNames=new HashSet<>(playerNames);
@@ -45,7 +54,7 @@ public class InputView {
         }
     }
 
-    private static void checkMaxLength(List<String> playerNames) {
+    private void checkMaxLength(List<String> playerNames) {
         for (String playerName: playerNames){
             if(playerName.length()>MAX_PLAYER_NAME_LENGTH){
                 throw new IllegalArgumentException("5글자 이하만 가능합니다");
